chat bot số 10/requirements.txt
streamlit==1.28.0
google-api-python-client==2.99.0
google-auth==2.22.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.1.0
PyPDF2==3.1.1
google-genai==1.0.0
